+++ CHANGELOG.md
@@ -7,11 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [2.0.0b2] - 2025-09-23
+## [2.0.0b3] - 2025-09-23
 
 ### Added
 
-- Added endpoint `get_flow_temperature_setpoint()` to get the calculated flow temperature setpoint 
+- Added endpoint `get_flow_temperature_setpoint()` to get the calculated flow temperature setpoint
+- Added endpoint `get_hot_water_flow()` to see if the fresh water module is activated and hot water flow
 
 ### Changed